+++ cli/cmd/status.go
@@ -17,6 +17,21 @@ func NewStatusCmd() *cobra.Command {
 	var statusCmd = &cobra.Command{
 		Use:   "status [<APP_NAME> | <APP_NAME:INSTANCE_NAME>]",
 		Short: "Status of the tarantool instance(s)",
+		Long: `The 'status' command provides information about the status of Tarantool instances.
+
+Columns:
+- INSTANCE: The name of the Tarantool instance.
+- STATUS: The current status of the instance:
+	- RUNNING: The instance is up and running.
+	- NOT RUNNING: The instance is not running.
+	- ERROR: The process has terminated unexpectedly.
+- PID: The watchdog process PID.
+- MODE: The mode of the instance, indicating its read/write status:
+	- RO: The instance is in read-only mode.
+	- RW: The instance is in read-write mode.
+- CONFIG: The config info status (for Tarantool 3+).
+- BOX: The box info status.
+- UPSTREAM: The replication upstream status.`,
 		Run: func(cmd *cobra.Command, args []string) {
 			cmdCtx.CommandName = cmd.Name()
 			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
@@ -35,6 +50,7 @@ func NewStatusCmd() *cobra.Command {
 	}
 
 	statusCmd.Flags().BoolVarP(&opts.Pretty, "pretty", "p", false, "pretty-print table")
+	statusCmd.Flags().BoolVarP(&opts.Details, "details", "d", false, "print detailed alerts.")
 
 	return statusCmd
 }